Copilot fixes

+++ src/main/java/com/sacristan/api/interfaces/admin/controllers/studentSequence/StudentSequenceController.java
@@ -62,15 +62,15 @@ public class StudentSequenceController {
     public ResponseEntity<List<SequenceResponse>> getUnassignedSequences(@PathVariable Long studentId){
         return ResponseEntity.ok(service.getUnassignedSequencesOfStudent(studentId).stream().map(SequenceResponse::of).toList());
     }
-    /*
+
     @Operation(
             summary = "Count assigned Sequences",
             description = "Count all the sequences assigned to a student"
     )
     @GetMapping("/{studentId}/sequence-count")
-    public ResponseEntity<Long> countAssignedSequences(@PathVariable Long studentId){
+    public ResponseEntity<Integer> countAssignedSequences(@PathVariable Long studentId){
         return ResponseEntity.ok(service.countAssignedSequencesOfStudent(studentId));
     }
-    */
+
 
 }
